refactor: update name of component
